# [Optimization] - Optimize key method