--- a/Neo.Cryptography.MPTTrie.Benchmarks/Neo.Cryptography.MPTTrie.Benchmarks.csproj
+++ b/Neo.Cryptography.MPTTrie.Benchmarks/Neo.Cryptography.MPTTrie.Benchmarks.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net9.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" Version="0.14.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\src\Neo\Neo.csproj" />
+  </ItemGroup>
+
+</Project>