# Description

Optimizes the Key method in Cache.cs by eliminating unnecessary stream abstractions and allocations during serialization of UInt256 values. A Benchmark has been added to validate the performance between both codes.


| Method        | Mean      | Error     | StdDev    | Median    |
|-------------- |----------:|----------:|----------:|----------:|
| Key_Original  | 27.236 ns | 0.5834 ns | 0.8367 ns | 27.691 ns |
| Key_Optimized |  4.864 ns | 0.0517 ns | 0.0458 ns |  4.856 ns |

Fixes # (issue)

## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Optimization (the change is only an optimization)
- [ ] Style (the change is only a code style for better maintenance or standard purpose)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# How Has This Been Tested?

<!-- Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration -->

- [x] Benchmarks_Cache

**Test Configuration**:


# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules